refactor: rename metric modules
