+++ backend/app/services/metrics/utils/math.py
